Add PIP_INDEX_URL env var to support custom index URL/mirror
Currently, Python S2I images use default pypi.python.org to get all
required packages during deployment process. However, there is a need
to allow users to use custom index URL (mirror) to get the packages.
This option can speed up deployment process.

The PIP_INDEX_URL env var which is already available in pip can be used
for that purposes. The information of PIP_INDEX_URL is added to the
README files to mark the availability of this option.

Signed-off-by: Vu Dinh <vdinh@redhat.com>