add buildpack metadata to 3.9 RHEL8 image
Similar to the work that was done in the Node.js images, this commit adds
buildpack metadata to the RHEL8 Dockerfile so that it can be used as a
"stack" when creating builder images. The OpenShift Serverless Functions
team is looking into using these images for our stacks, and these metadata
bits are a requirement to enable that.

See: https://github.com/sclorg/s2i-nodejs-container/pull/273

Signed-off-by: Lance Ball <lball@redhat.com>