exit oc4 tests gracefully  if login fails
The nounset shell option needs to be set after the OC login attempt,
because (from man shopt):

-u
  If expansion is attempted on an unset  variable  or
  parameter,  the  shell prints  an  error  message, and, if not
  interactive, exits with a non zero status.

That is in this matter unconvient, as we do not really know from the log
what has happened if the oc login fails. Moreover the testuite ends
with success, as only first failed test sets TESTSUITE_RESULT to 1.

If we check login with ct_os_check_login function, it handles the
situation conveniently.