Set VIRTUAL_ENV in venv_setup
If poetry is used to install dependencies in a Dockerfile, it will not
obey the virtualenv set up in the container because it does not detect
it properly. The dependencies end up installed into the system python
paths (eg. /usr/local/lib/python3.10/site-packages/).

This shows differing behaviour to pip which if used in the Dockerfile
to install dependencies will install them correctly into the virtual
env path (eg. /opt/app-root/lib/python3.10/site-packages/).

Setting VIRTUAL_ENV manually in the Dockerfile gives Poetry the needed
info to obey the virtual env paths.